--- lib/Command/src/Command.cpp
@@ -1,1 +1,0 @@
-#include "Command.h"
