--- lib/Command/src/Command.h
@@ -1,9 +1,0 @@
-#ifndef _LOG_COMMAND_H
-#define _LOG_COMMAND_H
-
-class Command {
-    public:
-        virtual void execute() = 0;
-};
-
-#endif
