+++ lib/Logger/src/CanLogger.cpp
@@ -1,0 +1,5 @@
+#include "CanLogger.h"
+
+CanLogger::CanLogger(SensorCan *sensor) {
+    this->_sensor = sensor;
+}
