+++ lib/Logger/src/GpsLogger.cpp
@@ -1,0 +1,5 @@
+#include "GpsLogger.h"
+
+GpsLogger::GpsLogger(SensorGps *sensor) {
+    this->_sensor = sensor;
+}
