+++ lib/Logger/src/ThermoLoggerrr.cpp
@@ -1,0 +1,7 @@
+#include "ThermoLogger.h"
+
+#include "settings.h"
+
+ThermoLogger::ThermoLogger(SensorThermo *sensor) {
+    this->_sensor = sensor;
+}
